Store the sequence number in StopSend() and resume it in StartSend().
When restarting the microphone device, we call StopSend() first, then StartSend() later. Since we reset sequence number in StopSend(), it sometimes causes libSRTP to complain about packets being replayed. Libjingle work around it by caching the sequence number in WebRtcVoiceEngine.cc, and call SetInitSequenceNumber() to resume the sequence number before StartSend().

This patch fixes this problem by storing the sequence number in StopSend(), and resume it in StartSend(). So that we can remove the workaround in libjingle.

BUG=2102
R=tommi@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1823004

git-svn-id: http://webrtc.googlecode.com/svn/stable@4374 4adac7df-926f-26a2-2b94-8c16560cd09d