Move use of DEPTH into build_with_chromium==1.
For build_with_chromium==1 the includes will be the same.
For build_with_chromium==0 the <(DEPTH) variable is replaced by ../..
which should be the same in all common use cases.

This change makes the include paths for all GYP targets
more similar to the setup in the
direct_dependent_settings section further down.

BUG=4185
TESTED=Trybots + build in Chromium with third_party/webrtc patched with this CL.
R=tommi@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/41739004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@8219 4adac7df-926f-26a2-2b94-8c16560cd09d